fix: keep plugin command llm lazy
