+++ src/agents/pi-embedded-runner/context-engine-capabilities.ts
@@ -1,15 +1,7 @@
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import type { ContextEngineRuntimeContext } from "../../context-engine/types.js";
-import {
-  parseAgentSessionKey,
-  normalizeAgentId,
-  normalizeMainKey,
-} from "../../routing/session-key.js";
-import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-} from "../../shared/string-coerce.js";
-import { resolveDefaultAgentId } from "../agent-scope.js";
+import { normalizeOptionalString } from "../../shared/string-coerce.js";
+import { resolveBoundAgentIdForSession } from "../session-agent-binding.js";
 
 export type ResolveContextEngineCapabilitiesParams = {
   config?: OpenClawConfig;
@@ -20,36 +12,6 @@ export type ResolveContextEngineCapabilitiesParams = {
   purpose: string;
 };
 
-function resolveBoundAgentId(params: {
-  config?: OpenClawConfig;
-  sessionKey?: string;
-  agentId?: string;
-}): string | undefined {
-  // Explicit agent ids are host-resolved at call sites that already know the
-  // active session agent, such as embedded attempts.
-  const explicitAgentId = normalizeOptionalString(params.agentId);
-  if (explicitAgentId) {
-    return normalizeAgentId(explicitAgentId);
-  }
-  // Canonical agent session keys carry the binding directly.
-  const normalizedSessionKey = normalizeOptionalString(params.sessionKey);
-  if (!normalizedSessionKey) {
-    return undefined;
-  }
-  const parsed = parseAgentSessionKey(normalizedSessionKey);
-  if (parsed?.agentId) {
-    return normalizeAgentId(parsed.agentId);
-  }
-  // Legacy main-session aliases are still active sessions; arbitrary legacy
-  // aliases stay unbound and fail closed in runtime LLM authorization.
-  const loweredSessionKey = normalizeLowercaseStringOrEmpty(normalizedSessionKey);
-  const mainKey = normalizeMainKey(params.config?.session?.mainKey);
-  if (loweredSessionKey === "main" || loweredSessionKey === mainKey) {
-    return resolveDefaultAgentId(params.config ?? {});
-  }
-  return undefined;
-}
-
 /**
  * Build host-owned capabilities that are bound to one context-engine runtime call.
  */
@@ -57,7 +19,7 @@ export function resolveContextEngineCapabilities(
   params: ResolveContextEngineCapabilitiesParams,
 ): Pick<ContextEngineRuntimeContext, "llm"> {
   const sessionKey = normalizeOptionalString(params.sessionKey);
-  const agentId = resolveBoundAgentId({
+  const agentId = resolveBoundAgentIdForSession({
     config: params.config,
     sessionKey,
     agentId: params.agentId,
